requirements.txt
python-dotenv
openai
numpy
